Revert "セッティング画面の作成"
