+++ src/containers/index.ts
@@ -6,4 +6,3 @@ export { default as UserEditScreen } from './UserEditScreen'
 export { default as PartyEntryScreen } from './PartyEntryScreen'
 export { default as HomeScreen } from './HomeScreen'
 export { default as PartyDetailScreen } from './PartyDetailScreen'
-export { default as SettingScreen } from './SettingScreen'
