+++ src/screens/index.ts
@@ -8,4 +8,3 @@ export { default as PostScreen } from './PostScreen'
 export { default as UserScreen } from './UserScreen'
 export { default as UserEditScreen } from './UserEditScreen'
 export { default as PartyDetailScreen } from './PartyDetailScreen'
-export { default as SettingScreen } from './SettingScreen'
